initial work on merge feeds for MTC
